AudioControlPython/requirements.txt
pyserial
pycaw
pywin32
